+++ src/Presentation/Nop.Web/Controllers/PrivateMessagesController.cs
@@ -1,17 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Web.Mvc;
 using Nop.Core;
 using Nop.Core.Domain.Customers;
 using Nop.Core.Domain.Forums;
 using Nop.Services.Customers;
 using Nop.Services.Forums;
-using Nop.Services.Helpers;
 using Nop.Services.Localization;
 using Nop.Services.Logging;
+using Nop.Web.Factories;
 using Nop.Web.Framework.Controllers;
 using Nop.Web.Framework.Security;
-using Nop.Web.Models.Common;
 using Nop.Web.Models.PrivateMessages;
 
 namespace Nop.Web.Controllers
@@ -21,43 +19,40 @@ namespace Nop.Web.Controllers
     {
         #region Fields
 
+        private readonly IPrivateMessagesModelFactory _privateMessagesModelFactory;
         private readonly IForumService _forumService;
         private readonly ICustomerService _customerService;
         private readonly ICustomerActivityService _customerActivityService;
         private readonly ILocalizationService _localizationService;
         private readonly IWorkContext _workContext;
         private readonly IStoreContext _storeContext;
-        private readonly IDateTimeHelper _dateTimeHelper;
         private readonly ForumSettings _forumSettings;
-        private readonly CustomerSettings _customerSettings;
 
         #endregion
 
         #region Constructors
 
-        public PrivateMessagesController(IForumService forumService,
-            ICustomerService customerService, ICustomerActivityService customerActivityService,
-            ILocalizationService localizationService, IWorkContext workContext, 
-            IStoreContext storeContext, IDateTimeHelper dateTimeHelper,
-            ForumSettings forumSettings, CustomerSettings customerSettings)
+        public PrivateMessagesController(IPrivateMessagesModelFactory privateMessagesModelFactory,
+            IForumService forumService,
+            ICustomerService customerService,
+            ICustomerActivityService customerActivityService,
+            ILocalizationService localizationService,
+            IWorkContext workContext, 
+            IStoreContext storeContext,
+            ForumSettings forumSettings)
         {
+            this._privateMessagesModelFactory = privateMessagesModelFactory;
             this._forumService = forumService;
             this._customerService = customerService;
             this._customerActivityService = customerActivityService;
             this._localizationService = localizationService;
             this._workContext = workContext;
             this._storeContext = storeContext;
-            this._dateTimeHelper = dateTimeHelper;
             this._forumSettings = forumSettings;
-            this._customerSettings = customerSettings;
         }
 
         #endregion
-
-        #region Utilities
-
-        #endregion
-
+        
         #region Methods
 
         public ActionResult Index(int? page, string tab)
@@ -72,36 +67,7 @@ namespace Nop.Web.Controllers
                 return new HttpUnauthorizedResult();
             }
 
-            int inboxPage = 0;
-            int sentItemsPage = 0;
-            bool sentItemsTabSelected = false;
-
-            switch (tab)
-            {
-                case "inbox":
-                    if (page.HasValue)
-                    {
-                        inboxPage = page.Value;
-                    }
-                    break;
-                case "sent":
-                    if (page.HasValue)
-                    {
-                        sentItemsPage = page.Value;
-                    }
-                    sentItemsTabSelected = true;
-                    break;
-                default:
-                    break;
-            }
-
-            var model = new PrivateMessageIndexModel
-            {
-                InboxPage = inboxPage,
-                SentItemsPage = sentItemsPage,
-                SentItemsTabSelected = sentItemsTabSelected
-            };
-
+            var model = _privateMessagesModelFactory.PreparePrivateMessageIndexModel(page, tab);
             return View(model);
         }
 
@@ -109,53 +75,7 @@ namespace Nop.Web.Controllers
         [ChildActionOnly]
         public ActionResult Inbox(int page, string tab)
         {
-            if (page > 0)
-            {
-                page -= 1;
-            }
-
-            var pageSize = _forumSettings.PrivateMessagesPageSize;
-
-            var list = _forumService.GetAllPrivateMessages(_storeContext.CurrentStore.Id,
-                0, _workContext.CurrentCustomer.Id, null, null, false, string.Empty, page, pageSize);
-
-            var inbox = new List<PrivateMessageModel>();
-
-            foreach (var pm in list)
-            {
-                inbox.Add(new PrivateMessageModel
-                {
-                    Id = pm.Id,
-                    FromCustomerId = pm.FromCustomer.Id,
-                    CustomerFromName = pm.FromCustomer.FormatUserName(),
-                    AllowViewingFromProfile = _customerSettings.AllowViewingProfiles && pm.FromCustomer != null && !pm.FromCustomer.IsGuest(),
-                    ToCustomerId = pm.ToCustomer.Id,
-                    CustomerToName = pm.ToCustomer.FormatUserName(),
-                    AllowViewingToProfile = _customerSettings.AllowViewingProfiles && pm.ToCustomer != null && !pm.ToCustomer.IsGuest(),
-                    Subject = pm.Subject,
-                    Message = pm.Text,
-                    CreatedOn = _dateTimeHelper.ConvertToUserTime( pm.CreatedOnUtc, DateTimeKind.Utc),
-                    IsRead = pm.IsRead,
-                });
-            }
-
-            var pagerModel = new PagerModel
-            {
-                PageSize = list.PageSize,
-                TotalRecords = list.TotalCount,
-                PageIndex = list.PageIndex,
-                ShowTotalSummary = false,
-                RouteActionName = "PrivateMessagesPaged",
-                UseRouteLinks = true,
-                RouteValues = new PrivateMessageRouteValues { page = page, tab = tab }
-            };
-
-            var model = new PrivateMessageListModel
-            {
-                Messages = inbox,
-                PagerModel = pagerModel
-            };
-
+            var model = _privateMessagesModelFactory.PrepareInboxModel(page, tab);
             return PartialView(model);
         }
 
@@ -163,53 +83,7 @@ namespace Nop.Web.Controllers
         [ChildActionOnly]
         public ActionResult SentItems(int page, string tab)
         {
-            if (page > 0)
-            {
-                page -= 1;
-            }
-
-            var pageSize = _forumSettings.PrivateMessagesPageSize;
-
-            var list = _forumService.GetAllPrivateMessages(_storeContext.CurrentStore.Id, 
-                _workContext.CurrentCustomer.Id, 0, null, false, null, string.Empty, page, pageSize);
-
-            var sentItems = new List<PrivateMessageModel>();
-
-            foreach (var pm in list)
-            {
-                sentItems.Add(new PrivateMessageModel
-                {
-                    Id = pm.Id,
-                    FromCustomerId = pm.FromCustomer.Id,
-                    CustomerFromName = pm.FromCustomer.FormatUserName(),
-                    AllowViewingFromProfile = _customerSettings.AllowViewingProfiles && pm.FromCustomer != null && !pm.FromCustomer.IsGuest(),
-                    ToCustomerId = pm.ToCustomer.Id,
-                    CustomerToName = pm.ToCustomer.FormatUserName(),
-                    AllowViewingToProfile = _customerSettings.AllowViewingProfiles && pm.ToCustomer != null && !pm.ToCustomer.IsGuest(),
-                    Subject = pm.Subject,
-                    Message = pm.Text,
-                    CreatedOn = _dateTimeHelper.ConvertToUserTime(pm.CreatedOnUtc, DateTimeKind.Utc),
-                    IsRead = pm.IsRead,
-                });
-            }
-
-            var pagerModel = new PagerModel
-            {
-                PageSize = list.PageSize,
-                TotalRecords = list.TotalCount,
-                PageIndex = list.PageIndex,
-                ShowTotalSummary = false,
-                RouteActionName = "PrivateMessagesPaged",
-                UseRouteLinks = true,
-                RouteValues = new PrivateMessageRouteValues { page = page, tab = tab }
-            };
-
-            var model = new PrivateMessageListModel
-            {
-                Messages = sentItems,
-                PagerModel = pagerModel
-            };
-
+            var model = _privateMessagesModelFactory.PrepareSentModel(page, tab);
             return PartialView(model);
         }
 
@@ -305,46 +179,23 @@ namespace Nop.Web.Controllers
         public ActionResult SendPM(int toCustomerId, int? replyToMessageId)
         {
             if (!_forumSettings.AllowPrivateMessages)
-            {
                 return RedirectToRoute("HomePage");
-            }
 
             if (_workContext.CurrentCustomer.IsGuest())
-            {
                 return new HttpUnauthorizedResult();
-            }
 
             var customerTo = _customerService.GetCustomerById(toCustomerId);
-
             if (customerTo == null || customerTo.IsGuest())
-            {
                 return RedirectToRoute("PrivateMessages");
-            }
 
-            var model = new SendPrivateMessageModel();
-            model.ToCustomerId = customerTo.Id;
-            model.CustomerToName = customerTo.FormatUserName();
-            model.AllowViewingToProfile = _customerSettings.AllowViewingProfiles && !customerTo.IsGuest();
-
+            PrivateMessage replyToPM = null;
             if (replyToMessageId.HasValue)
             {
                 //reply to a previous PM
-                var replyToPM = _forumService.GetPrivateMessageById(replyToMessageId.Value);
-                if (replyToPM == null)
-                {
-                    return RedirectToRoute("PrivateMessages");
-                }
-
-                if (replyToPM.ToCustomerId == _workContext.CurrentCustomer.Id || replyToPM.FromCustomerId == _workContext.CurrentCustomer.Id)
-                {
-                    model.ReplyToMessageId = replyToPM.Id;
-                    model.Subject = string.Format("Re: {0}", replyToPM.Subject);
-                }
-                else
-                {
-                    return RedirectToRoute("PrivateMessages");
-                }
+                replyToPM = _forumService.GetPrivateMessageById(replyToMessageId.Value);
             }
+
+            var model = _privateMessagesModelFactory.PrepareSendPrivateMessageModel(customerTo, replyToPM);
             return View(model);
         }
 
@@ -389,9 +240,6 @@ namespace Nop.Web.Controllers
             {
                 return RedirectToRoute("PrivateMessages");
             }
-            model.ToCustomerId = toCustomer.Id;
-            model.CustomerToName = toCustomer.FormatUserName();
-            model.AllowViewingToProfile = _customerSettings.AllowViewingProfiles && !toCustomer.IsGuest();
 
             if (ModelState.IsValid)
             {
@@ -437,6 +285,7 @@ namespace Nop.Web.Controllers
                 }
             }
 
+            model = _privateMessagesModelFactory.PrepareSendPrivateMessageModel(toCustomer, replyToPM);
             return View(model);
         }
 
@@ -471,21 +320,7 @@ namespace Nop.Web.Controllers
                 return RedirectToRoute("PrivateMessages");
             }
 
-            var model = new PrivateMessageModel
-            {
-                Id = pm.Id,
-                FromCustomerId = pm.FromCustomer.Id,
-                CustomerFromName = pm.FromCustomer.FormatUserName(),
-                AllowViewingFromProfile = _customerSettings.AllowViewingProfiles && pm.FromCustomer != null && !pm.FromCustomer.IsGuest(),
-                ToCustomerId = pm.ToCustomer.Id,
-                CustomerToName = pm.ToCustomer.FormatUserName(),
-                AllowViewingToProfile = _customerSettings.AllowViewingProfiles && pm.ToCustomer != null && !pm.ToCustomer.IsGuest(),
-                Subject = pm.Subject,
-                Message = pm.FormatPrivateMessageText(),
-                CreatedOn = _dateTimeHelper.ConvertToUserTime(pm.CreatedOnUtc, DateTimeKind.Utc),
-                IsRead = pm.IsRead,
-            };
-
+            var model = _privateMessagesModelFactory.PreparePrivateMessageModel(pm);
             return View(model);
         }
