+++ src/Presentation/Nop.Web/Factories/PrivateMessagesModelFactory.cs
@@ -1,0 +1,235 @@
+﻿using System;
+using System.Collections.Generic;
+using Nop.Core;
+using Nop.Core.Domain.Customers;
+using Nop.Core.Domain.Forums;
+using Nop.Services.Customers;
+using Nop.Services.Forums;
+using Nop.Services.Helpers;
+using Nop.Web.Models.Common;
+using Nop.Web.Models.PrivateMessages;
+
+namespace Nop.Web.Factories
+{
+    public partial class PrivateMessagesModelFactory : IPrivateMessagesModelFactory
+    {
+        #region Fields
+
+        private readonly IForumService _forumService;
+        private readonly IWorkContext _workContext;
+        private readonly IStoreContext _storeContext;
+        private readonly IDateTimeHelper _dateTimeHelper;
+        private readonly ForumSettings _forumSettings;
+        private readonly CustomerSettings _customerSettings;
+
+        #endregion
+
+        #region Constructors
+
+        public PrivateMessagesModelFactory(IForumService forumService,
+            IWorkContext workContext, 
+            IStoreContext storeContext,
+            IDateTimeHelper dateTimeHelper,
+            ForumSettings forumSettings,
+            CustomerSettings customerSettings)
+        {
+            this._forumService = forumService;
+            this._workContext = workContext;
+            this._storeContext = storeContext;
+            this._dateTimeHelper = dateTimeHelper;
+            this._forumSettings = forumSettings;
+            this._customerSettings = customerSettings;
+        }
+
+        #endregion
+
+        #region Methods
+
+        public virtual PrivateMessageIndexModel PreparePrivateMessageIndexModel(int? page, string tab)
+        {
+            int inboxPage = 0;
+            int sentItemsPage = 0;
+            bool sentItemsTabSelected = false;
+
+            switch (tab)
+            {
+                case "inbox":
+                    if (page.HasValue)
+                    {
+                        inboxPage = page.Value;
+                    }
+                    break;
+                case "sent":
+                    if (page.HasValue)
+                    {
+                        sentItemsPage = page.Value;
+                    }
+                    sentItemsTabSelected = true;
+                    break;
+                default:
+                    break;
+            }
+
+            var model = new PrivateMessageIndexModel
+            {
+                InboxPage = inboxPage,
+                SentItemsPage = sentItemsPage,
+                SentItemsTabSelected = sentItemsTabSelected
+            };
+
+            return model;
+        }
+
+        public virtual PrivateMessageListModel PrepareInboxModel(int page, string tab)
+        {
+            if (page > 0)
+            {
+                page -= 1;
+            }
+
+            var pageSize = _forumSettings.PrivateMessagesPageSize;
+
+            var list = _forumService.GetAllPrivateMessages(_storeContext.CurrentStore.Id,
+                0, _workContext.CurrentCustomer.Id, null, null, false, string.Empty, page, pageSize);
+
+            var inbox = new List<PrivateMessageModel>();
+
+            foreach (var pm in list)
+            {
+                inbox.Add(new PrivateMessageModel
+                {
+                    Id = pm.Id,
+                    FromCustomerId = pm.FromCustomer.Id,
+                    CustomerFromName = pm.FromCustomer.FormatUserName(),
+                    AllowViewingFromProfile = _customerSettings.AllowViewingProfiles && pm.FromCustomer != null && !pm.FromCustomer.IsGuest(),
+                    ToCustomerId = pm.ToCustomer.Id,
+                    CustomerToName = pm.ToCustomer.FormatUserName(),
+                    AllowViewingToProfile = _customerSettings.AllowViewingProfiles && pm.ToCustomer != null && !pm.ToCustomer.IsGuest(),
+                    Subject = pm.Subject,
+                    Message = pm.Text,
+                    CreatedOn = _dateTimeHelper.ConvertToUserTime( pm.CreatedOnUtc, DateTimeKind.Utc),
+                    IsRead = pm.IsRead,
+                });
+            }
+
+            var pagerModel = new PagerModel
+            {
+                PageSize = list.PageSize,
+                TotalRecords = list.TotalCount,
+                PageIndex = list.PageIndex,
+                ShowTotalSummary = false,
+                RouteActionName = "PrivateMessagesPaged",
+                UseRouteLinks = true,
+                RouteValues = new PrivateMessageRouteValues { page = page, tab = tab }
+            };
+
+            var model = new PrivateMessageListModel
+            {
+                Messages = inbox,
+                PagerModel = pagerModel
+            };
+
+            return model;
+        }
+
+        public virtual PrivateMessageListModel PrepareSentModel(int page, string tab)
+        {
+            if (page > 0)
+            {
+                page -= 1;
+            }
+
+            var pageSize = _forumSettings.PrivateMessagesPageSize;
+
+            var list = _forumService.GetAllPrivateMessages(_storeContext.CurrentStore.Id, 
+                _workContext.CurrentCustomer.Id, 0, null, false, null, string.Empty, page, pageSize);
+
+            var sentItems = new List<PrivateMessageModel>();
+
+            foreach (var pm in list)
+            {
+                sentItems.Add(new PrivateMessageModel
+                {
+                    Id = pm.Id,
+                    FromCustomerId = pm.FromCustomer.Id,
+                    CustomerFromName = pm.FromCustomer.FormatUserName(),
+                    AllowViewingFromProfile = _customerSettings.AllowViewingProfiles && pm.FromCustomer != null && !pm.FromCustomer.IsGuest(),
+                    ToCustomerId = pm.ToCustomer.Id,
+                    CustomerToName = pm.ToCustomer.FormatUserName(),
+                    AllowViewingToProfile = _customerSettings.AllowViewingProfiles && pm.ToCustomer != null && !pm.ToCustomer.IsGuest(),
+                    Subject = pm.Subject,
+                    Message = pm.Text,
+                    CreatedOn = _dateTimeHelper.ConvertToUserTime(pm.CreatedOnUtc, DateTimeKind.Utc),
+                    IsRead = pm.IsRead,
+                });
+            }
+
+            var pagerModel = new PagerModel
+            {
+                PageSize = list.PageSize,
+                TotalRecords = list.TotalCount,
+                PageIndex = list.PageIndex,
+                ShowTotalSummary = false,
+                RouteActionName = "PrivateMessagesPaged",
+                UseRouteLinks = true,
+                RouteValues = new PrivateMessageRouteValues { page = page, tab = tab }
+            };
+
+            var model = new PrivateMessageListModel
+            {
+                Messages = sentItems,
+                PagerModel = pagerModel
+            };
+
+            return model;
+        }
+
+        public virtual SendPrivateMessageModel PrepareSendPrivateMessageModel(Customer customerTo, PrivateMessage replyToPM)
+        {
+            if (customerTo == null)
+                throw new ArgumentNullException("customerTo");
+
+            var model = new SendPrivateMessageModel();
+            model.ToCustomerId = customerTo.Id;
+            model.CustomerToName = customerTo.FormatUserName();
+            model.AllowViewingToProfile = _customerSettings.AllowViewingProfiles && !customerTo.IsGuest();
+
+            if (replyToPM == null)
+                return model;
+
+            if (replyToPM.ToCustomerId == _workContext.CurrentCustomer.Id ||
+                replyToPM.FromCustomerId == _workContext.CurrentCustomer.Id)
+            {
+                model.ReplyToMessageId = replyToPM.Id;
+                model.Subject = string.Format("Re: {0}", replyToPM.Subject);
+            }
+
+            return model;
+        }
+
+        public virtual PrivateMessageModel PreparePrivateMessageModel(PrivateMessage pm)
+        {
+            if (pm == null)
+                throw new ArgumentNullException("pm");
+
+            var model = new PrivateMessageModel
+            {
+                Id = pm.Id,
+                FromCustomerId = pm.FromCustomer.Id,
+                CustomerFromName = pm.FromCustomer.FormatUserName(),
+                AllowViewingFromProfile = _customerSettings.AllowViewingProfiles && pm.FromCustomer != null && !pm.FromCustomer.IsGuest(),
+                ToCustomerId = pm.ToCustomer.Id,
+                CustomerToName = pm.ToCustomer.FormatUserName(),
+                AllowViewingToProfile = _customerSettings.AllowViewingProfiles && pm.ToCustomer != null && !pm.ToCustomer.IsGuest(),
+                Subject = pm.Subject,
+                Message = pm.FormatPrivateMessageText(),
+                CreatedOn = _dateTimeHelper.ConvertToUserTime(pm.CreatedOnUtc, DateTimeKind.Utc),
+                IsRead = pm.IsRead,
+            };
+
+            return model;
+        }
+
+        #endregion
+    }
+}
