+++ src/Presentation/Nop.Web/Nop.Web.csproj
@@ -281,6 +281,8 @@
     <Compile Include="Factories\IOrderModelFactory.cs" />
     <Compile Include="Factories\OrderModelFactory.cs" />
     <Compile Include="Factories\IProfileModelFactory.cs" />
+    <Compile Include="Factories\IPrivateMessagesModelFactory.cs" />
+    <Compile Include="Factories\PrivateMessagesModelFactory.cs" />
     <Compile Include="Factories\ProfileModelFactory.cs" />
     <Compile Include="Factories\IReturnRequestModelFactory.cs" />
     <Compile Include="Factories\ReturnRequestModelFactory.cs" />
